luna_soc/sim: move sim.py to util/sim.py
